+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v1.2.1
 * Add Google Nexus 7 (2013)
 * #13: Fix the width of device sub-menu
 * #15: Set the Ajax loading indicator at the center of screen
+* #16: Cannot move the mouse out of the frame after resizing the frame
 
 v1.2.0 (2013/09/27)